feat: use local svg
