+++ package.json
@@ -25,6 +25,7 @@
     "gatsby": "^4.0.0",
     "gatsby-plugin-layout": "^3.4.0",
     "gatsby-plugin-mdx": "^3.4.0",
+    "gatsby-plugin-react-svg": "^3.1.0",
     "js-search": "^2.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
@@ -38,8 +39,7 @@
     "http-proxy-middleware": "^2.0.1",
     "postcss": "^8.3.7",
     "prettier": "^2.5.1",
-    "tailwindcss": "^3.0.0",
-    "typescript": "^4.5.4"
+    "tailwindcss": "^3.0.0"
   },
   "repository": {
     "type": "git",
