--- resource/icons/mirror.tsx
@@ -1,25 +1,0 @@
-import React, { Component } from "react"
-
-export default (props: { size: string }) => (
-  <svg viewBox="0 0 1024 1024" version="1.1"
-       xmlns="http://www.w3.org/2000/svg" width={props.size} height={props.size} className="stroke-current">
-    <path
-      d="M403.29 775.26H172.9a73.44 73.44 0 0 1-73.44-73.44v-532.2a73.44 73.44 0 0 1 73.44-73.45h230.39a73.44 73.44 0 0 1 73.44 73.45v532.2a73.44 73.44 0 0 1-73.44 73.44z m-2-75.46V171.63H174.92V699.8z"
-    />
-    <path
-      d="M99.46 850.71m37.73 0l301.81 0q37.73 0 37.73 37.73l0-0.01q0 37.73-37.73 37.73l-301.81 0q-37.73 0-37.73-37.73l0 0.01q0-37.73 37.73-37.73Z"
-    />
-    <path
-      d="M552.18 850.71m37.73 0l301.81 0q37.73 0 37.73 37.73l0-0.01q0 37.73-37.73 37.73l-301.81 0q-37.73 0-37.73-37.73l0 0.01q0-37.73 37.73-37.73Z"
-    />
-    <path
-      d="M856 775.26H625.63a73.45 73.45 0 0 1-73.45-73.44v-532.2a73.45 73.45 0 0 1 73.45-73.45H856a73.44 73.44 0 0 1 73.44 73.45v532.2A73.44 73.44 0 0 1 856 775.26z m-2-75.46V171.63H627.64V699.8z"
-    />
-    <path
-      d="M250.37 284.81m33.74 0l7.97 0q33.74 0 33.74 33.74l0 234.33q0 33.74-33.74 33.74l-7.97 0q-33.74 0-33.74-33.74l0-234.33q0-33.74 33.74-33.74Z"
-    />
-    <path
-      d="M703.09 284.81m33.74 0l7.97 0q33.74 0 33.74 33.74l0 234.33q0 33.74-33.74 33.74l-7.97 0q-33.74 0-33.74-33.74l0-234.33q0-33.74 33.74-33.74Z"
-    />
-  </svg>
-)
